added input validation service and implemented it within service classes

+++ CreditCardRewardTracker/src/com/revature/creditcardrewardtracker/service/AdminService.java
@@ -9,12 +9,16 @@ import com.revature.creditcardrewardtracker.dao.UserRepoDB;
 
 public class AdminService {
 	
-	Scanner sc;	
-	IUserRepo d;
+	private Scanner sc;	
+	private IUserRepo d;
+	private InputValidationService inputValidation;
+	private ValidationService validation;
 	
 	public AdminService(Scanner sc, Connection connection) {
 		this.sc = sc;
 		d = new UserRepoDB(connection);
+		inputValidation = new InputValidationService(sc);
+		validation = new ValidationService(connection);
 	}
 	
 	public List<String> getAllUsers() {
@@ -25,57 +29,80 @@ public class AdminService {
 	
 	public void changeUserPassword() {
 		System.out.println("Please enter the username associated with the account to reset password.");
-		String username = sc.next();
+		String username = inputValidation.getValidStringInput();
+		boolean usernameExists = validation.usernameExistsValidation(username);
 	
-		System.out.println("Please enter new password for user.");
-		String newPassword = sc.next();
-
-		boolean result = d.changePassword(username, newPassword);
-		if (result == true) {
-			System.out.println(username + " password's updated.");
+		if (usernameExists == true) {
+			System.out.println("Please enter new password for user.");
+			String newPassword = inputValidation.getValidStringInput();
+			boolean isValidPassword = validation.passwordLengthValidation(newPassword);
+			
+			if (isValidPassword == true) {
+				boolean result = d.changePassword(username, newPassword);
+				if (result == true) {
+					System.out.println(username + " password's updated.");
+				} else {
+					System.out.println(username + " password not changed.");
+				}
+			}
 		} else {
-			System.out.println(username + " password not changed.");
+			System.out.println("User " + username + " not found.");
 		}
 	}
 	
 	public void deleteUserAccount() {
-		System.out.println("Please enter the username associated with the account to be delted.");
-		String username = sc.next();
+		System.out.println("Please enter the username associated with the account to be deleted.");
+		String username = inputValidation.getValidStringInput();
+		boolean usernameExists = validation.usernameExistsValidation(username);
 		
-		boolean result = false;
-		System.out.println("Enter YES to verify account deletion for " + username);
-		if (sc.next().equalsIgnoreCase("yes")) {
-			result = d.deleteUser(username);
-		}
-		
-		if (result == true ) {
-			System.out.println(username + "'s account deleted.");
+		if (usernameExists == true) {
+			boolean result = false;
+			System.out.println("Enter YES to verify account deletion for " + username);
+			if (sc.next().equalsIgnoreCase("yes")) {
+				result = d.deleteUser(username);
+			}
+			
+			if (result == true ) {
+				System.out.println(username + "'s account deleted.");
+			} else {
+				System.out.println(username + "'s account not deleted.");
+			}
 		} else {
-			System.out.println(username + "'s account not deleted.");
+			System.out.println("User " + username + " not found.");
 		}
 	}
 	
 	public void promoteUserAccount() {
 		System.out.println("Please enter the username associated with the account to be promoted to admin.");
-		String username = sc.next();
+		String username = inputValidation.getValidStringInput();
+		boolean usernameExists = validation.usernameExistsValidation(username);
 		
-		boolean result = d.promoteAdmin(username);
-		if (result == true) {
-			System.out.println(username + " promoted to admin successfully.");
-		} else {
-			System.out.println(username + " not promoted to admin.");
+		if(usernameExists == true) {
+			boolean result = d.promoteAdmin(username);
+			if (result == true) {
+				System.out.println(username + " promoted to admin successfully.");
+			} else {
+				System.out.println(username + " not promoted to admin.");
+			}
+		}else {
+			System.out.println("User " + username + " not found.");
 		}
 	}
 	
 	public void demoteUserAccount() {
 		System.out.println("Please enter the username associated with the account to be demoted to user.");
-		String username = sc.next();
+		String username = inputValidation.getValidStringInput();
+		boolean usernameExists = validation.usernameExistsValidation(username);
 		
-		boolean result = d.demoteAdmin(username);
-		if (result == true) {
-			System.out.println(username + " demoted to standard account successfully.");
+		if (usernameExists == true) {
+			boolean result = d.demoteAdmin(username);
+			if (result == true) {
+				System.out.println(username + " demoted to standard account successfully.");
+			} else {
+				System.out.println(username + " not demoted to standard account.");
+			}
 		} else {
-			System.out.println(username + " not demoted to standard account.");
+			System.out.println("User " + username + " not found.");
 		}
 	}
